task 6

+++ 0x00-hello_world/6-size.c
@@ -16,11 +16,11 @@ int main(void)
 	long int li;
 	long long int lli;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned int)sizeof(c));
-	printf("Size of an int: %lu byte(s) \n", (unsigned int)sizeof(i));
-	printf("Size of a long int: %lu byte(s) \n", (unsigned int)sizeof(li));
-	printf("Size of a long long int: %lu byte(s) \n", (unsigned int)sizeof(lli));
-	printf("Size of a float: %lu byte(s) \n", (unsigned int)sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s) \n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s) \n", (unsigned long)sizeof(li));
+	printf("Size of a long long int: %lu byte(s) \n", (unsigned long)sizeof(lli));
+	printf("Size of a float: %lu byte(s) \n", (unsigned long)sizeof(f));
 	return (0);
 
 }
